Add notification uris to ApiAppDao
